Ironically fixed a bug that revealed a backup bug in OpenWRT
sync-led was an improper procd init script.  It didn't include
wrapper for the main code.

This actually revealed a bug in OpenWRT:
[sysupgrade -b produces invalid tar.gz file](https://github.com/openwrt/openwrt/issues/14931)

It's a interesting bug.  There are a ton of network-based and
host-based virus scanners and intrusion detection systems out there
that could choke on a carefully crafted file, or worse.

The default ash shell on OpenWRT supports the -e and -n command line
options, so crafting arbitrary byte sequences is possible.